requirements.txt
Flask
gTTS
python-docx
PyPDF2